be/ref: refactoring integer id to long

+++ back-end/src/main/java/com/ssafy/eoullim/repository/AnimonRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface AnimonRepository extends JpaRepository<AnimonEntity, Integer> {
+public interface AnimonRepository extends JpaRepository<AnimonEntity, Long> {
     @Query(value = "SELECT A FROM AnimonEntity A WHERE A.id BETWEEN 1 and 4")
     List<AnimonEntity> getDefaultAnimon();
